add fixes in ms checkout.

+++ ms-checkout/src/main/java/shop/style/checkout/DTO/Form/CartFormDTO.java
@@ -5,7 +5,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Positive;
 
 @Data
 @AllArgsConstructor
@@ -16,5 +15,5 @@ public class CartFormDTO {
     private String variant_id;
 
     @NotNull
-    private Long quantity;
+    private Integer quantity;
 }
